reusing old code to calculate densities
