+++ tools/ppls1/ppls1/fluids/ljts/ljts.py
@@ -6,20 +6,28 @@ Created on Sun May 31 12:32:59 2020
 """
 
 #%% Get saturated densities by Vrabec et al., Molecular Physics 104 (2006).
-def rho_vrabec2006(T):
+def rho_vrabec2006(t,Re=None):
     '''
-    Get saturated densities by Vrabec et al., Molecular Physics 104 (2006)
+    Get saturated densities by Vrabec et al., Molecular Physics 104 (2006). Equation numbers refer this paper.
 
-    :param float T: Temperature
+    :param float T: Temperature, float Re: reduced radius of spherical interface
     :return: float rhol, float rhov: Saturated liquid and vapor density
     '''
     
     tc,rc=1.0779,0.3190
     dt=tc-t
     a,b,c=0.5649,0.1314,0.0413
-    rhol=rc+a*dt**(1/3.)+b*dt+c*dt**(3/2.)
+    rhol=rc+a*dt**(1/3.)+b*dt+c*dt**(3/2.)                      # equation 4
     a,b,c=0.5649,0.2128,0.0702
-    rhov=rc-a*dt**(1/3.)+b*dt+c*dt**(3/2.)
+    rhov=rc-a*dt**(1/3.)+b*dt+c*dt**(3/2.)                      # equation 5
+
+    if Re != None:
+        a,b,c,d,e=0.1485,0.0471,1.272,0.4944,5.0090
+        rhol=rhol+(a-b*(tc-t)**2)/(Re+c-d*(tc-t)**-1+e*(tc-t))  # equation 18
+        
+        a,b,c,d,e=0.0541,0.1682,1.0810,0.2490,10.87
+        rhov=rhov+(-a+b*t**2)/(Re-c-d*(tc-t)**-1+e*(tc-t))      # equation 19
+    
     return rhol,rhov
 
 if __name__ == "__main__":
